+++ tests/test_export/test_base.py
@@ -4,7 +4,7 @@ from django.core.exceptions import FieldDoesNotExist
 from django.db import models
 from import_export.fields import Field
 
-from dbentry.export.base import MIZResource
+from dbentry.export.base import MIZResource, get_verbose_name_for_resource_field
 from dbentry.export.fields import AnnotationField
 from dbentry.export.widgets import YesNoBooleanWidget
 from tests.case import MIZTestCase
@@ -23,7 +23,6 @@ class DummyResource(MIZResource):
 
 class TestMIZResource(MIZTestCase):
     def setUp(self):
-        super().setUp()
         self.resource = DummyResource()
 
     def test_filter_export_calls_defer_fts(self):
@@ -111,45 +110,15 @@ class TestMIZResource(MIZTestCase):
 
     def test_get_export_headers(self):
         """
-        Assert that get_export_headers prefers the value given by the
-        column_name argument if column_name isn't the default value (which
-        would be equal to the value of the attribute argument).
+        Assert that get_export_headers calls get_verbose_name_for_resource_field
+        for every export field.
         """
-        export_fields = [Field(attribute="name_set", column_name="bar")]
-        with patch.object(self.resource, "get_export_fields", new=Mock(return_value=export_fields)):
-            self.assertEqual(self.resource.get_export_headers(), ["bar"])
-
-    def test_get_export_headers_model_field_verbose_name(self):
-        """
-        Assert that get_export_headers uses the field's verbose name if
-        column_name has the default value (equal to attribute argument).
-        """
-        export_fields = [Field(attribute="name_set", column_name="name_set")]
-        with patch.object(self.resource, "get_export_fields", new=Mock(return_value=export_fields)):
-            self.assertEqual(self.resource.get_export_headers(), ["Verbose Name Explicitly Set"])
-
-    def test_get_export_headers_not_a_model_field(self):
-        export_fields = [Field(attribute="not_on_model", column_name="A Different Field")]
-        with patch.object(self.resource, "get_export_fields", new=Mock(return_value=export_fields)):
-            assert self.resource.get_export_headers() == ["A Different Field"]
-
-    def test_get_export_headers_column_name_is_none(self):
-        """
-        Assert that get_export_headers uses the field's verbose name if the
-        value for the column_name argument is None.
-        """
-        export_fields = [Field(attribute="name_set", column_name=None)]
-        with patch.object(self.resource, "get_export_fields", new=Mock(return_value=export_fields)):
-            self.assertEqual(self.resource.get_export_headers(), ["Verbose Name Explicitly Set"])
-
-    def test_get_export_headers_capitalizes_derived_verbose_name(self):
-        """
-        Assert that get_export_headers capitalizes a verbose name that has been
-        derived from the model field's name.
-        """
-        export_fields = [Field(attribute="no_name_set")]
-        with patch.object(self.resource, "get_export_fields", new=Mock(return_value=export_fields)):
-            self.assertEqual(self.resource.get_export_headers(), ["No name set"])
+        export_fields = [Mock(), Mock()]
+        with patch("dbentry.export.base.get_verbose_name_for_resource_field") as get_verbose_mock:
+            with patch.object(self.resource, "get_export_fields", new=Mock(return_value=export_fields)):
+                with patch.object(self.resource, "get_field_name"):
+                    self.resource.get_export_headers()
+        self.assertEqual(get_verbose_mock.call_count, len(export_fields))
 
     def test_get_annotations(self):
         export_fields = ["any", AnnotationField(attribute="foo", expr="bar")]
@@ -167,3 +136,62 @@ class TestMIZResource(MIZTestCase):
     def test_widget_from_django_field_no_boolean_field(self):
         field = models.IntegerField()
         self.assertNotEqual(MIZResource.widget_from_django_field(field), YesNoBooleanWidget)
+
+
+class TestGetVerboseName(MIZTestCase):
+    def setUp(self):
+        self.resource = DummyResource()
+
+    def test_get_verbose_name_for_resource_field_column_name_set(self):
+        """
+        Assert that get_verbose_name_for_resource_field prioritizes an
+        explicitly set column_name over other names.
+        """
+        field = Field(attribute="name_set", column_name="Explicit Column Name")
+        with patch.object(self.resource, "get_export_fields", new=Mock(return_value=[field])):
+            with patch.object(self.resource, "fields", new={"name_set": field}):
+                self.assertEqual(get_verbose_name_for_resource_field(self.resource, "name_set"), "Explicit Column Name")
+
+    def test_get_verbose_name_for_resource_field_model_verbose_name(self):
+        """
+        Assert that get_verbose_name_for_resource_field uses the verbose name
+        defined on the model field if no explicit column_name is set.
+        """
+        field = Field(attribute="name_set", column_name="name_set")
+        with patch.object(self.resource, "get_export_fields", new=Mock(return_value=[field])):
+            with patch.object(self.resource, "fields", new={"name_set": field}):
+                self.assertEqual(
+                    get_verbose_name_for_resource_field(self.resource, "name_set"), "Verbose Name Explicitly Set"
+                )
+
+    def test_get_verbose_name_for_resource_field_derived_verbose_name(self):
+        """
+        Assert that get_verbose_name_for_resource_field capitalizes the first
+        letter of a model field verbose name that has been derived from the
+        model field's name because no explicit verbose name was set.
+        """
+        field = Field(attribute="no_name_set")
+        with patch.object(self.resource, "get_export_fields", new=Mock(return_value=[field])):
+            with patch.object(self.resource, "fields", new={"no_name_set": field}):
+                self.assertEqual(get_verbose_name_for_resource_field(self.resource, "no_name_set"), "No name set")
+
+    def test_get_verbose_name_for_resource_field_no_model_field_column_name(self):
+        """
+        Assert that get_verbose_name_for_resource_field uses the field's name
+        if no model field could be found and no explicit column name was set.
+        """
+        field = Field(attribute="does_not_exist")
+        with patch.object(self.resource, "get_export_fields", new=Mock(return_value=[field])):
+            with patch.object(self.resource, "fields", new={"does_not_exist": field}):
+                self.assertEqual(get_verbose_name_for_resource_field(self.resource, "does_not_exist"), "Does not exist")
+
+    def test_get_verbose_name_for_resource_field_no_resource_field(self):
+        """
+        Assert that get_verbose_name_for_resource_field raises a ValueError if
+        the resource does not have a field with the given name.
+        """
+
+        def test_callable():
+            return get_verbose_name_for_resource_field(self.resource, "foo")
+
+        self.assertRaises(ValueError, test_callable)